Facility: change facility to be event sourced.
Taken from our ShoppingCart example.
Note: Postman does not fully work at this point.

Added ResourceEntity, modified Resource.
Resource now has a time window to book on and a size for this window. A window is just an array of time slots, like hours in a day.
There is also a nowPointer, which points in the array to which slot we are on now.