Facility: rename Api to Mod, because it is the model part of the API.
I don't know if i can call this domain or not. In which case, the model is what I am calling now `domain`.
Because 'domain' refers to the objects that are known to their domain experts, while 'model' could be anything that programmers come up with to work better with code.