Facility: return the API objects from getters.
Previously they were returning the domain objects, not the API.
Note that the same could be done for `ReservationEntity`'s `@GetMapping`, but `ReservationState` is read-only and quite informative.
Eventually it should be a View, but for now, I will leave it as API.
There could be a concept of reservation response object, which will be API (in `Mod`), and the domain's `ReservationState` could be left alone as private.