Facility: wip with StubAction, DelegatingServiceAction.
Added CommandSourcedEntity...
This represents the first attempt in dealing with commands to Google Calendar that could fail.
If a command fails, we don't want to lose it, we want to be able to repeat the operation.
For that, we can possibly think of using a command-sourced entity.
For all this to work, Google Calendar calls must be idempotent.
There should be one `CommandSourcedEntity` per `FacilityEntity`.