Facility: implement confirmation message in twist thread, add ACL.
Removed CommandSourcedEntity: none is needed, because the projection in DelegatingServiceAction is good enough.
If the call to the external service fails, as long as the result of that is in the returned effect, the projection will be retried.
And as long as the saveToGoogle call is idempotent, that is ok.

Also upgraded kalix-sdk.version.

Next todos:
- deploy into kalix with real url
- pass the correct reservation details in the message
- check idempotency of REST calls
- double check domain-wide auth with Google
- check if effects().asyncReply is good