Workflow: reservation.
Introduce Workflow as a reservation coordinator in place of ReservationEntity and its related events and actions.

Note that in a workflow implementation, actions depend on workflow, unlike with event-wiring where actions depend on entities.
Note that this implementation of reservation entity (workflow) has time outs, so that when the reservation is passed, the workflow can no longer be exercised (COMPLETE).

Workflow: renaming in ReservationEntity.

Renames and simplifications to align to master's branch:
- dropped dto postfix
- Booked -> Fulfilled
- InitiateReservation -> Init
- BookingRejected -> ReservationRejected
- InquireBooking -> CheckAvailability
- BookingAccepted -> ReservationAccepted
- BookingRejected -> ReservationRejected
- BookingCanceled -> ReservationCanceled