Undo all experimental and troubleshooting changes.
This is the tip of a broken branch only useful to keep history.
This branch will be merged to master along with the 'good' branch containing good changes.

The broken branch keeps no changes in the merge. The broken branch shows the problems encountered while changing ResourceState with Jackson errors and an obscure infinite loop to later analyze.

Note: the broken branch was based on the implementation of ReservationEntity as a Workflow, while the good branch (a couple of commits only) contains changes on top of the implementation of ReservationEntity as a regular entity. All Workflow related changes are removed in this merge, as well as the changes that made the integ test to fail. ResourceState works in this merge now.