Reservation: implement Reservation to use broadcast.
The previous implementation of reservation was going through the list of resources available one by one until one was found available.
This implementation instead broadcasts the reservation request to all resources and the first replying that it is available is selected for booking.
The ReservationEntity does not know how many resources will respond to the broadcast, and so, it will have to wait a maximum time before giving up.
If between the time the reservation is notified of availability to when it actually tries to seal the booking, the booking is no longer possible, then the reservation is tried on the next resource that notified of being available (ReservationEntity keeps track of that in a availability Set).

In this commit:
- `FacilityAction` initiate the reservation and timer associated to it
- `ReservationAction` does the broadcast.
- `ReservationState` now takes a `Set` of available resources. These are the resources entities that replied 'yes' to the broadcast request of availability.
- `ReservationEntity` now has an additional state, COLLECTING, in which the `ReservationEntity` collects the responses to the broadcast from the various (unknown) resource entities.
- As soon as `ReservationEntity` collects a yes from a `ResourceEntity`, it goes into SELECTING state, from which it will proceed by booking the resource if still possible.