Reservation app refactorings: List to Set for collecting resource ids, upgrade and rename all ancillary packages.
Had to change from List to Set for resource ids, which makes more sense.
As a result of that change, lots of dependencies had to be upgraded or changed, like all modules dependent on the spi module (EventDetails API changed as a result of that).
Package names were also wrong inside supporting packages, so I corrected all those.