Customer: move customer app inside reservation.
The entire Customer code could be, like it was up until now, a different service.
However, it complicates deployments for this demo, and so, to simplify, in this commit I moved that code inside `reservation` so that I have only one deployment to make and test, just one service.
All the endpoints of Facility, of User and of Asset entities are now deployable along with Reservation, Resource and Pool entities.
Note: the name "asset" here refers to a "resource", which is the entity that knows about being reservable or not. A Facility knows about its resources (= its assets), but will not know what is available or not in terms of reserving: it has to ask its children.
A Pool entity is also a container of reservable stuff, but it is more general than a Facility, as it can contain end resources (=assets, leaf entities) or containers (other facilities).