Customer module: customer endpoints don't work, errors.
I here try to fix this customer module dependency of reservation so that it gets picked up, but it does not work.
ComponentScan does not find customer: no customer endpoints are available.
I am attempting here to change package names in maven to improve the hierarchy and also renaming 'resource' to 'asset' so to ease collisions with reservation's 'resource'. I will ask for help on this, so it is good to present some code.