Facility: correct asset registration and renames.
From FacilityEntity, you can now either create and register an asset in it in one call, or you can first create an asset with AssetEntity in one call, and then register it with FacilityEntity in a second call.
Both ways will go through the same code path when it comes to the registration of the asset with the facility.
Other than correcting asset registration, I aligned names (events, commands, methods) and also added validation when creating assets and facilities.