Reservation: bug found, corrected initialization of entities.
The bug to be fixed is about the selection of ids to be sent to google calendar.
It is minor, but it has to be fixed, or else the calendar is not going to display correctly in google (the drop down of calendars to choose from).
There is a todo in code for this.

When it comes to the initialization of entities, it can be done better than it is now, by trying to distinguish between its first empty state and after initialization: you do not want to re-initialize an entity by overriding its content by mistake.
That should be prevented from happening.